Ensure file mode is correct when reading or writting HBA (WIP)
The "hba" module works with native string type ("str") internally (there is no
unicode literals). This means that, when performing IO operations, we need to
operate in binary mode in Python 2 and text mode in Python 3.

As this is not quite obvious, we add a check on file mode when calling .parse()
and .save().

Added test demonstrates how it works.

WIP: some tests need update.